Reorganize as core Sentry module with registration pattern
Restructure codebase to support multiple modules logging to same logbook:

- Create ILogBookEventHandler interface for module registration
- Move Sentry events to Modules/Sentry/Events/
- Move SentryRepository to Modules/Sentry/
- Move LogBookEventHandler to Modules/Sentry/EventHandlers/ as SentryLogBookEventHandler
- SentryLogBookEventHandler now implements ILogBookEventHandler
- Update Program.cs to auto-discover and initialize all ILogBookEventHandler implementations
- Move localization resources to match new handler class name

Sentry is now a complete, self-contained module that can be paralleled by
other modules (Patrol, Maintenance, etc.) without modifying core infrastructure.
Each module independently registers its events and logbook handlers.

New modules follow the pattern:
1. Create Modules/{ModuleName}/ structure
2. Define events in Events/
3. Create handler implementing ILogBookEventHandler in EventHandlers/
4. Register in DI as ILogBookEventHandler
5. Program.cs automatically discovers and initializes all handlers
