+++ src/GraphQL/DigitalGuardBookQuery.cs
@@ -1,5 +1,6 @@
 using DigitalGuardBook.GraphQL.Types;
 using DigitalGuardBook.Repositories;
+using DigitalGuardBook.Modules.Sentry;
 using GraphQL;
 using GraphQL.Types;
