+++ src/Infrastructure/ILogBookEventHandler.cs
@@ -1,0 +1,6 @@
+namespace DigitalGuardBook.Infrastructure;
+
+public interface ILogBookEventHandler
+{
+    void Register(InProcessEventPublisher publisher);
+}
