+++ src/Modules/Sentry/Events/GuardServiceEndedEvent.cs
@@ -1,4 +1,4 @@
-namespace DigitalGuardBook.Events;
+namespace DigitalGuardBook.Modules.Sentry.Events;
 
 public sealed record GuardServiceEndedEvent(
     string PersonId,
