+++ src/Modules/Sentry/Events/GuardServiceStartedEvent.cs
@@ -1,4 +1,4 @@
-namespace DigitalGuardBook.Events;
+namespace DigitalGuardBook.Modules.Sentry.Events;
 
 public sealed record GuardServiceStartedEvent(
     string PersonId,
