+++ src/Modules/Sentry/Events/SentryFinishedEvent.cs
@@ -1,4 +1,4 @@
-namespace DigitalGuardBook.Events;
+namespace DigitalGuardBook.Modules.Sentry.Events;
 
 public sealed record SentryFinishedEvent(
     DateTimeOffset FinishTime,
