+++ src/Modules/Sentry/Events/SentryStartedEvent.cs
@@ -1,4 +1,4 @@
-namespace DigitalGuardBook.Events;
+namespace DigitalGuardBook.Modules.Sentry.Events;
 
 public sealed record SentryStartedEvent(
     DateTimeOffset StartTime,
