+++ src/Modules/Sentry/Events/SupervisorServiceEndedEvent.cs
@@ -1,4 +1,4 @@
-namespace DigitalGuardBook.Events;
+namespace DigitalGuardBook.Modules.Sentry.Events;
 
 public sealed record SupervisorServiceEndedEvent(
     string PersonId,
