+++ src/Modules/Sentry/Events/SupervisorServiceStartedEvent.cs
@@ -1,4 +1,4 @@
-namespace DigitalGuardBook.Events;
+namespace DigitalGuardBook.Modules.Sentry.Events;
 
 public sealed record SupervisorServiceStartedEvent(
     string PersonId,
