+++ src/Resources/EventHandlers/SentryLogBookEventHandler.de-DE.resx
